Drop `Channel`, move functionality to `Client`
