+++ lib/transport.js
@@ -1,7 +1,7 @@
 'use strict';
 
+const http = require('node:http');
 const metautil = require('metautil');
-const { Channel } = require('./channel.js');
 
 const MIME_TYPES = {
   bin: 'application/octet-stream',
@@ -83,9 +83,36 @@ const LOCATION = 'Path=/; Domain';
 const COOKIE_DELETE = `${TOKEN}=deleted; Expires=${EPOCH}; ${LOCATION}=`;
 const COOKIE_HOST = `Expires=${FUTURE}; ${LOCATION}`;
 
-class HttpChannel extends Channel {
-  constructor(server, req, res) {
-    super(server, req, res);
+class Transport {
+  constructor(console, req) {
+    this.console = console;
+    this.req = req;
+    this.ip = req.socket.remoteAddress;
+  }
+
+  error(code = 500, { callId, error = null, httpCode = null } = {}) {
+    const { console, req, ip } = this;
+    const { url, method } = req;
+    if (!httpCode) httpCode = (error && error.httpCode) || code;
+    const status = http.STATUS_CODES[httpCode];
+    const pass = httpCode < 500 || httpCode > 599;
+    const message = pass && error ? error.message : status || 'Unknown error';
+    const reason = `${httpCode}\t${code}\t${error ? error.stack : status}`;
+    console.error(`${ip}\t${method}\t${url}\t${reason}`);
+    const packet = { callback: callId, error: { message, code } };
+    this.send(packet, httpCode);
+  }
+
+  send(obj, code = 200) {
+    const data = JSON.stringify(obj);
+    this.write(data, code, 'json');
+  }
+}
+
+class HttpTransport extends Transport {
+  constructor(console, req, res) {
+    super(console, req);
+    this.res = res;
     res.on('close', () => {
       this.destroy();
     });
@@ -97,11 +124,6 @@ class HttpChannel extends Channel {
     const mimeType = MIME_TYPES[ext] || MIME_TYPES.html;
     res.writeHead(httpCode, { ...HEADERS, 'Content-Type': mimeType });
     res.end(data);
-  }
-
-  send(obj, httpCode = 200) {
-    const data = JSON.stringify(obj);
-    this.write(data, httpCode, 'json');
   }
 
   redirect(location) {
@@ -151,11 +173,22 @@ class HttpChannel extends Channel {
   }
 }
 
-const createChannel = (server, req, res) => new HttpChannel(server, req, res);
+class WsTransport extends Transport {
+  constructor(console, req, connection) {
+    super(console, req);
+    this.connection = connection;
+    connection.on('message', (data, isBinary) => {
+      if (isBinary) this.binary(data);
+      else this.message(data);
+    });
+    connection.on('close', () => {
+      this.destroy();
+    });
+  }
 
-const addHeaders = (headers) => {
-  const { origin } = headers;
-  if (origin) HEADERS['Access-Control-Allow-Origin'] = origin;
-};
+  send(obj) {
+    this.write(JSON.stringify(obj));
+  }
+}
 
-module.exports = { createChannel, addHeaders };
+module.exports = { Transport, HttpTransport, WsTransport, MIME_TYPES, HEADERS };
