+++ metacom.js
@@ -2,8 +2,5 @@
 
 const { Metacom } = require('./lib/client.js');
 const { Server } = require('./lib/server.js');
-const { Channel } = require('./lib/channel.js');
 
-module.exports.Metacom = Metacom;
-module.exports.Server = Server;
-module.exports.Channel = Channel;
+module.exports = { Metacom, Server };
